+++ packages/localize/CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - It is no longer necessary to provide a message `id`. When omitted, an id will
   be automatically generated from the string contents of the template.
 
+- Remove unintentional export of `_msg` function (use `msg`).
+
 ## [0.5.1] - 2020-11-09
 
 ### Fixed